requirements.txt
opencv-python
cmake
pyyaml
streamlit
face_recognition
pandas
pickle

